requirements.txt
openai
anthropic
google-genai
google-generativeai
boto3
pyyaml
